+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.29.0
 ### Feature
-* add animation to exec [#679](https://github.com/jcubic/jquery.terminal/issues/679)
+* add animation to `exec` and `execAnimation` option [#679](https://github.com/jcubic/jquery.terminal/issues/679)
 ### Bugfix
 * tweak types for cmd function prompt